requirements-dev.txt
pytest==6.2.5
aws_cdk.integ_tests_alpha